fix: add checkstyle and fix it
