+++ src/main/java/lilybot/LilyBotException.java
@@ -3,7 +3,7 @@ package lilybot;
 /**
  * Aims to handle exceptions.
  */
-public class LilyBotException extends Exception{
+public class LilyBotException extends Exception {
     public LilyBotException(String warn) {
         super(warn);
     }
