+++ src/main/java/lilybot/Main.java
@@ -1,7 +1,5 @@
 package lilybot;
 
-import lilybot.Gui.MainWindow;
-
 import java.io.IOException;
 
 import javafx.application.Application;
@@ -9,7 +7,8 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
-import lilybot.Storage.Storage;
+import lilybot.gui.MainWindow;
+import lilybot.storage.Storage;
 
 //@@author SE-EDU
 //Reused from https://se-education.org/guides/tutorials/javaFx.html
