+++ src/main/java/lilybot/command/ListCommand.java
@@ -1,0 +1,41 @@
+package lilybot.command;
+
+import lilybot.gui.Ui;
+import lilybot.task.TaskList;
+
+/**
+ * Command for listing all tasks.
+ *
+ */
+public class ListCommand implements Command {
+
+    private Ui ui;
+    private String command;
+    private TaskList taskList;
+
+    /**
+     * Constructs ListCommand with the following constructor.
+     *
+     * @param ui To be displayed for users.
+     * @param command Command entered by users.
+     * @param taskList For tracking the list of tasks.
+     */
+    public ListCommand(Ui ui, String command, TaskList taskList) {
+        this.ui = ui;
+        this.command = command;
+        this.taskList = taskList;
+    }
+
+    /**
+     * Lists the tasks.
+     *
+     * @param ui To be displayed for users.
+     * @param command Command entered by users.
+     * @param taskList For tracking the list of tasks.
+     * @return The messages to be displayed after execution.
+     */
+    @Override
+    public String exceute(Ui ui, String command, TaskList taskList) {
+        return ui.listTask(taskList);
+    }
+}
